% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{analysis_summary}
\alias{analysis_summary}
\title{Analysis summary}
\usage{
analysis_summary(metrics_table)
}
\arguments{
\item{metrics_table}{Output from \code{calculate_metrics()}
or \code{convert_table_unit()}.}
}
\description{
prepare table summaries
}
