% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_tools.R
\name{convert_image_size_unit}
\alias{convert_image_size_unit}
\title{Convert image sizes}
\usage{
convert_image_size_unit(image_path, width = NULL, height = NULL, dpi = NULL)
}
\arguments{
\item{image_path}{Path to image.}

\item{width}{Image width in centimeter.}

\item{height}{Image height in centimeter.}

\item{dpi}{Dots Per Inch unit used for conversion.}
}
\description{
Convert image sizes to pixel from other units.
}
