% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{convert_table_unit}
\alias{convert_table_unit}
\title{Convert Unit used by Metrics Table}
\usage{
convert_table_unit(
  metrics_table,
  conversion_rate_width = NULL,
  conversion_rate_height = NULL,
  unit = "cm"
)
}
\arguments{
\item{metrics_table}{Output from `calculate_metrics()`.}

\item{conversion_rate_width}{Default: NULL. Value of unit conversion rate.
If rate values differ from width and height, a length two numeric vector
with conversion rates for width and lengths can be supplied.}

\item{conversion_rate_height}{Default: NULL. Value of unit conversion rate.
If rate values differ from width and height, a length two numeric vector
with conversion rates for width and lengths can be supplied.}

\item{unit}{Unit to convert values from pixel.}
}
\description{
Convert units used to calculate metrics.
}
