% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_tools.R
\name{convert_video_to_image}
\alias{convert_video_to_image}
\title{Convert video to image}
\usage{
convert_video_to_image(video_path, frames_path, fps = 5)
}
\arguments{
\item{video_path}{Path to video file.}

\item{frames_path}{Path to save frames extracted from video.}

\item{fps}{default = 5; Frames per second to decompose video.}
}
\description{
Convert video to image
}
\seealso{
Other video_tools: 
\code{\link{extract_video_info}()},
\code{\link{proccess_video}()}
}
\concept{video_tools}
