% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{count_area_circle}
\alias{count_area_circle}
\title{Count occurrences in circular area}
\usage{
count_area_circle(dist_table, diameter_px = 50)
}
\arguments{
\item{dist_table}{Table containing position of the mass center of the object
per frame.}

\item{diameter_px}{Area side length measured in pixels.}
}
\description{
Calculates how many times other points
  happens in an area around the main point.
}
