% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{count_area_square}
\alias{count_area_square}
\title{Count occurrences in square area}
\usage{
count_area_square(dist_table, side_px = 50)
}
\arguments{
\item{dist_table}{Table containing position of the mass center of the object
per frame.}

\item{side_px}{Area side length measured in pixels.}
}
\description{
Calculates how many times other points
  happens in an area around the main point.
}
