% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphazis-package.R
\docType{package}
\name{emphazis-package}
\alias{emphazis}
\alias{emphazis-package}
\title{emphazis: Emphazis}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
