% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{extract_video_info}
\alias{extract_video_info}
\title{Convert video to image}
\usage{
extract_video_info(video_path)
}
\arguments{
\item{video_path}{Path to video file.}
}
\description{
Convert video to image
}
\seealso{
Other video_process: 
\code{\link{convert_video_to_image}()},
\code{\link{proccess_video}()}
}
\concept{video_process}
