% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{generate_subject_model}
\alias{generate_subject_model}
\title{Generate subject model}
\usage{
generate_subject_model(subject_path, background_path)
}
\arguments{
\item{subject_path}{Path to subject image.}

\item{background_path}{Path to background image.}
}
\description{
Generate subject model.
}
