% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_average_speed}
\alias{plot_average_speed}
\title{Plot average speed
Plot cumulative sum of distance traveled by the subject.}
\usage{
plot_average_speed(dist_table, range = NULL)
}
\description{
Plot average speed
Plot cumulative sum of distance traveled by the subject.
}
