% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cumulative_distance}
\alias{plot_cumulative_distance}
\title{Plot Cumulative distance
Plot cumulative sum of distance traveled by the subject.}
\usage{
plot_cumulative_distance(dist_table, range = NULL)
}
\description{
Plot Cumulative distance
Plot cumulative sum of distance traveled by the subject.
}
