% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{prepare_path_data}
\alias{prepare_path_data}
\title{Prepare path data}
\usage{
prepare_path_data(res_list)
}
\arguments{
\item{res_list}{Output from video processing.}
}
\description{
Prepare data for path plotting.
}
