% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{proccess_video}
\alias{proccess_video}
\title{Extract data from video}
\usage{
proccess_video(
  video_path,
  frames_path,
  subject_model,
  coord1 = c(285, 655),
  coord2 = c(475, 20)
)
}
\arguments{
\item{frames_path}{Path to save frames extracted from video.}

\item{subject_model}{model generated from `generate_subject_model()`.}

\item{coord1}{length 2 vector with position of the top right point.}

\item{coord2}{length 2 vector with position of the bottom left point.}
}
\description{
Extract data and frames from video.
}
