% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rate_to_unit}
\alias{rate_to_unit}
\title{Get measurements}
\usage{
rate_to_unit(image_path, rate, type)
}
\arguments{
\item{image_path}{image path.}

\item{rate}{conversion rate.}

\item{type}{measurement type. "width" or "height".}
}
\description{
Get measurement values from pixel and conversion rate.
}
