% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_tools.R
\name{slice_image}
\alias{slice_image}
\title{Slice image}
\usage{
slice_image(image_path, output_path, coord1 = NULL, coord2 = NULL)
}
\arguments{
\item{image_path}{Path to image file.}

\item{output_path}{Path to save sliced image.}

\item{coord1}{Length 2 vector with position of the top left point.}

\item{coord2}{Length 2 vector with position of the bottom right point.}
}
\description{
Slice image with given X and Y coordinates.
}
\concept{image_tools}
