% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{analysis_summary}
\alias{analysis_summary}
\title{Analysis summary}
\usage{
analysis_summary(metrics_table, unit = "cm")
}
\arguments{
\item{metrics_table}{Output from `calculate_metrics()`.}

\item{unit}{Unit to convert values from pixel.}
}
\description{
prepare table summaries
}
