% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{analysis_summary}
\alias{analysis_summary}
\title{Analysis summary}
\usage{
analysis_summary(dist_table)
}
\arguments{
\item{dist_table}{Output from `calculate_distances`.}
}
\description{
prepare table summaries
}
