% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker.R
\name{calculate_distances}
\alias{calculate_distances}
\title{Calculate Distances}
\usage{
calculate_distances(res_list)
}
\arguments{
\item{res_list}{Output from video processing.}
}
\description{
Calculate distances and cumulative distances from frame to frame.
}
