% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Distances}
\usage{
calculate_metrics(res_df, fps = 5, conversion_rate = NULL)
}
\arguments{
\item{res_df}{Output from video processing.}

\item{conversion_rate}{Default: NULL. Value of unit conversion rate.
If rate values differ from width and height, a length two numeric vector
with conversion rates for width and lengths can be supplied.}
}
\description{
Calculate distances and cumulative distances from frame to frame.
}
