% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_metrics.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Metrics}
\usage{
calculate_metrics(position_table)
}
\arguments{
\item{position_table}{Coordinate position of subject per frame.}
}
\value{
tibble with additional attributes.
}
\description{
Calculate distances traveled, cumulative distances, time,
  and other metrics from frame to frame.
}
