% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{count_area_circle}
\alias{count_area_circle}
\title{Count occurrences in circular area}
\usage{
count_area_circle(metrics_table, diameter_pct = 5)
}
\arguments{
\item{metrics_table}{Table containing position of the mass center of the object
per frame.}

\item{diameter_pct}{Circular area diameter around object.}
}
\description{
Calculates how many times other points
happens in an area around the main point.
}
