% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emphazis-package.R
\docType{package}
\name{emphazis-package}
\alias{emphazis}
\alias{emphazis-package}
\title{emphazis: Emphazis - Zebrafish tracking made simple}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Computational vision tracking tailored for zebrafish
    tracking studies.
}
\keyword{internal}
