% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_tools.R
\name{proccess_video}
\alias{proccess_video}
\title{Extract data from video}
\usage{
proccess_video(
  video_path,
  subject_model,
  frames_path,
  coord1 = NULL,
  coord2 = NULL,
  fps = 5
)
}
\arguments{
\item{video_path}{Path to video file.}

\item{subject_model}{model generated from `generate_subject_model()`.}

\item{frames_path}{Path to save frames extracted from video.}

\item{coord1}{length 2 vector with position of the top left point.}

\item{coord2}{length 2 vector with position of the bottom right point.}

\item{fps}{default = 5; Frames per second to decompose video.}
}
\description{
Extract data and frames from video.
}
\seealso{
Other video_tools: 
\code{\link{convert_video_to_image}()},
\code{\link{extract_video_info}()}
}
\concept{video_tools}
